+++ modules/sdk-lib-mpc/test/unit/tss/ecdsa/paillierproof.ts
@@ -3,7 +3,7 @@ import should from 'should';
 import { alpha, m, generateP, prove, verify } from '../../../../src/tss/ecdsa/paillierproof';
 import { hexToBigInt, minModulusBitLength, randomBigInt } from '../../../../src';
 import { deserializePaillierChallenge, deserializePaillierChallengeProofs } from '../../../../src/tss/ecdsa/types';
-import { mockedPaillierProofs } from '../../../paillierproof.util';
+import { mockedPaillierProofs } from '../../../fixtures/paillierpProof.util';
 import { primesSmallerThan319567 } from '../../../../src/tss/ecdsa/primes';
 
 describe('EcdsaPaillierProof', function () {
